Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	long long int long_long_int_type;
 	float float_type;
 
-	printf("Size of char: %ld byte(s)\n", sizeof(char_type));
-	printf("Size of int: %ld byte(s)\n", sizeof(int_type));
-	printf("Size of long int: %ld byte(s)\n", sizeof(long_int_type));
-	printf("Size of long long int: %ld bytes(s)\n", sizeof(long_long_int_type));
-	printf("Size of float: %ld byte(s)\n", sizeof(float_type));
+	printf("Size of char: %ld byte(s)\n",(unsigned long)sizeof(char_type));
+	printf("Size of int: %ld byte(s)\n", (unsigned long)sizeof(int_type));
+	printf("Size of long int: %ld byte(s)\n", (unsigned long)sizeof(long_int_type));
+	printf("Size of long long int: %ld bytes(s)\n", (unsigned long)sizeof(long_long_int_type));
+	printf("Size of float: %ld byte(s)\n", (unsigned long)sizeof(float_type));
 	return (0);
 }
